Minimum Spanning Tree (MSP) | Kruskal's Algorithm
A minimum spanning tree is an exclusive sort of length (“weight”) oriented tree that minimizes the lengths of the edges of the tree. It's generally used for network designs, find approximate solutions for complex mathematical problems, etc. I tried to explain briefly with Kruskal's Algorithm and a fundamental example | Exciting Future that Lies Ahead